Extracted out public api types to their own file
