+++ src/index.ts
@@ -2,7 +2,11 @@ import puppeteerConfig from './puppeteer-config.json';
 
 export { default as createScraper } from './scrapers/factory';
 export { SCRAPERS, CompanyTypes } from './definitions';
-export { ScraperOptions } from './scrapers/base-scraper';
+
+// Note: the typo ScaperScrapingResult is extracted here for backward compatibility
+export {
+  ScraperOptions, ScraperScrapingResult as ScaperScrapingResult, ScraperScrapingResult,
+} from './scrapers/interface';
 
 export function getPuppeteerConfig() {
   return { ...puppeteerConfig };
