+++ src/scrapers/base-scraper.ts
@@ -1,20 +1,13 @@
 import { EventEmitter } from 'events';
-import { Browser, Page } from 'puppeteer';
 import moment from 'moment-timezone';
 import { TimeoutError } from '../helpers/waiting';
-import { TransactionsAccount } from '../transactions';
-import { CompanyTypes } from '../definitions';
+import type { ScraperErrorTypes } from './errors';
+import { createGenericError, createTimeoutError } from './errors';
+import {
+  ScraperScrapingResult, ScraperCredentials, ScraperOptions, Scraper,
+} from './interface';
 
 const SCRAPE_PROGRESS = 'SCRAPE_PROGRESS';
-
-export enum ScraperErrorTypes {
-  InvalidPassword ='INVALID_PASSWORD',
-  ChangePassword = 'CHANGE_PASSWORD',
-  Timeout = 'TIMEOUT',
-  AccountBlocked = 'ACCOUNT_BLOCKED',
-  Generic = 'GENERIC',
-  General = 'GENERAL_ERROR'
-}
 
 export interface ScaperLoginResult {
   success: boolean;
@@ -22,125 +15,7 @@ export interface ScaperLoginResult {
   errorMessage?: string; // only on success=false
 }
 
-export interface FutureDebit {
-  amount: number;
-  amountCurrency: string;
-  chargeDate?: string;
-  bankAccountNumber?: string;
-}
-
-export interface ScaperScrapingResult {
-  success: boolean;
-  accounts?: TransactionsAccount[];
-  futureDebits?: FutureDebit[];
-  errorType?: ScraperErrorTypes;
-  errorMessage?: string; // only on success=false
-}
-
-export type ScraperCredentials = Record<string, string>;
-
-export interface ScraperOptions {
-  /**
-   * The company you want to scrape
-   */
-  companyId: CompanyTypes;
-
-  /**
-   * include more debug info about in the output
-   */
-  verbose?: boolean;
-
-  /**
-   * the date to fetch transactions from (can't be before the minimum allowed time difference for the scraper)
-   */
-  startDate: Date;
-
-  /**
-   * shows the browser while scraping, good for debugging (default false)
-   */
-  showBrowser?: boolean;
-
-
-  /**
-   * scrape transactions to be processed X months in the future
-   */
-  futureMonthsToScrape?: number;
-
-  /**
-   * option from init puppeteer browser instance outside the libary scope. you can get
-   * browser diretly from puppeteer via `puppeteer.launch()`
-   */
-  browser?: any;
-
-  /**
-   * provide a patch to local chromium to be used by puppeteer. Relevant when using
-   * `israeli-bank-scrapers-core` library
-   */
-  executablePath?: string;
-
-  /**
-   * if set to true, all installment transactions will be combine into the first one
-   */
-  combineInstallments?: boolean;
-
-  /**
-   * additional arguments to pass to the browser instance. The list of flags can be found in
-   *
-   * https://developer.mozilla.org/en-US/docs/Mozilla/Command_Line_Options
-   * https://peter.sh/experiments/chromium-command-line-switches/
-   */
-  args?: string[];
-
-  /**
-   * Maximum navigation time in milliseconds, pass 0 to disable timeout.
-   * @default 30000
-   */
-  timeout?: number | undefined;
-
-  /**
-   * adjust the browser instance before it is being used
-   *
-   * @param browser
-   */
-  prepareBrowser?: (browser: Browser) => Promise<void>;
-
-  /**
-   * adjust the page instance before it is being used.
-   *
-   * @param page
-   */
-  preparePage?: (page: Page) => Promise<void>;
-
-  /**
-   * if set, store a screenshot if failed to scrape. Used for debug purposes
-   */
-  storeFailureScreenShotPath?: string;
-
-  /**
-   * if set, will set the timeout in milliseconds of puppeteer's `page.setDefaultTimeout`.
-   */
-  defaultTimeout?: number;
-
-  /**
-   * Options for manipulation of output data
-   */
-  outputData?: OutputDataOptions;
-
-  /**
-   * Perform additional operation for each transaction to get more information (Like category) about it.
-   * Please note: It will take more time to finish the process.
-   */
-  additionalTransactionInformation?: boolean;
-}
-
-export interface OutputDataOptions {
-  /**
-   * if true, the result wouldn't be filtered out by date, and you will return unfiltered scrapped data.
-   */
-  enableTransactionsFilterByDate?: boolean;
-}
-
-export enum ScaperProgressTypes {
+export enum ScraperProgressTypes {
   Initializing = 'INITIALIZING',
   StartScraping = 'START_SCRAPING',
   LoggingIn = 'LOGGING_IN',
@@ -151,23 +26,7 @@ export enum ScaperProgressTypes {
   Terminating = 'TERMINATING',
 }
 
-function createErrorResult(errorType: ScraperErrorTypes, errorMessage: string) {
-  return {
-    success: false,
-    errorType,
-    errorMessage,
-  };
-}
-
-function createTimeoutError(errorMessage: string) {
-  return createErrorResult(ScraperErrorTypes.Timeout, errorMessage);
-}
-
-function createGenericError(errorMessage: string) {
-  return createErrorResult(ScraperErrorTypes.Generic, errorMessage);
-}
-
-export class BaseScraper {
+export class BaseScraper implements Scraper {
   private eventEmitter = new EventEmitter();
 
   constructor(public options: ScraperOptions) {
@@ -175,12 +34,12 @@ export class BaseScraper {
 
   // eslint-disable-next-line  @typescript-eslint/require-await
   async initialize() {
-    this.emitProgress(ScaperProgressTypes.Initializing);
+    this.emitProgress(ScraperProgressTypes.Initializing);
     moment.tz.setDefault('Asia/Jerusalem');
   }
 
-  async scrape(credentials: ScraperCredentials): Promise<ScaperScrapingResult> {
-    this.emitProgress(ScaperProgressTypes.StartScraping);
+  async scrape(credentials: ScraperCredentials): Promise<ScraperScrapingResult> {
+    this.emitProgress(ScraperProgressTypes.StartScraping);
     await this.initialize();
 
     let loginResult;
@@ -212,7 +71,7 @@ export class BaseScraper {
     } catch (e) {
       scrapeResult = createGenericError(e.message);
     }
-    this.emitProgress(ScaperProgressTypes.EndScraping);
+    this.emitProgress(ScraperProgressTypes.EndScraping);
 
     return scrapeResult;
   }
@@ -223,16 +82,16 @@ export class BaseScraper {
   }
 
   // eslint-disable-next-line  @typescript-eslint/require-await
-  async fetchData(): Promise<ScaperScrapingResult> {
+  async fetchData(): Promise<ScraperScrapingResult> {
     throw new Error(`fetchData() is not created in ${this.options.companyId}`);
   }
 
   // eslint-disable-next-line @typescript-eslint/no-unused-vars, @typescript-eslint/require-await
   async terminate(_success: boolean) {
-    this.emitProgress(ScaperProgressTypes.Terminating);
+    this.emitProgress(ScraperProgressTypes.Terminating);
   }
 
-  emitProgress(type: ScaperProgressTypes) {
+  emitProgress(type: ScraperProgressTypes) {
     this.emit(SCRAPE_PROGRESS, { type });
   }
