+++ src/scrapers/errors.ts
@@ -1,0 +1,30 @@
+export enum ScraperErrorTypes {
+  InvalidPassword = 'INVALID_PASSWORD',
+  ChangePassword = 'CHANGE_PASSWORD',
+  Timeout = 'TIMEOUT',
+  AccountBlocked = 'ACCOUNT_BLOCKED',
+  Generic = 'GENERIC',
+  General = 'GENERAL_ERROR'
+}
+
+export type ErrorResult = {
+  success: false;
+  errorType: ScraperErrorTypes;
+  errorMessage: string;
+};
+
+function createErrorResult(errorType: ScraperErrorTypes, errorMessage: string): ErrorResult {
+  return {
+    success: false,
+    errorType,
+    errorMessage,
+  };
+}
+
+export function createTimeoutError(errorMessage: string): ErrorResult {
+  return createErrorResult(ScraperErrorTypes.Timeout, errorMessage);
+}
+
+export function createGenericError(errorMessage: string): ErrorResult {
+  return createErrorResult(ScraperErrorTypes.Generic, errorMessage);
+}
