+++ src/scrapers/max.ts
@@ -8,8 +8,8 @@ import { waitUntilElementFound, elementPresentOnPage, clickButton } from '../hel
 import getAllMonthMoments from '../helpers/dates';
 import { fixInstallments, sortTransactionsByDate, filterOldTransactions } from '../helpers/transactions';
 import { Transaction, TransactionStatuses, TransactionTypes } from '../transactions';
-import { ScraperOptions, ScraperCredentials } from './base-scraper';
 import { getDebug } from '../helpers/debug';
+import { ScraperCredentials, ScraperOptions } from './interface';
 
 const debug = getDebug('max');
